move merge function under internal/util package
